requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pandas>=2.2.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
python-pptx>=1.0.2,<2.0.0
python-multipart>=0.0.9,<1.0.0
httpx>=0.27.0,<1.0.0
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
pypdf>=5.0.0,<6.0.0
